# docs/swagger.yaml
basePath: /
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  model.BasicInfo:
    properties:
      birthday:
        type: string
      createdAt:
        type: string
      custom_info:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      gender:
        type: integer
      id:
        type: integer
      marriage:
        type: integer
      name:
        type: string
      nation:
        type: string
      native_place:
        type: string
      phone_number:
        type: string
      political_status:
        type: integer
      profile_pic:
        type: string
      resume_id:
        type: string
      updatedAt:
        type: string
      work_experience:
        type: string
    type: object
  model.Resume:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      resume_id:
        type: string
      resume_name:
        type: string
      updatedAt:
        type: string
      user_id:
        type: string
    type: object
  request.BasicInfo:
    properties:
      birthday:
        type: string
      custom_info:
        type: string
      email:
        type: string
      gender:
        description: required 字段不能为空和初始值
        type: integer
      marriage:
        type: integer
      name:
        type: string
      nation:
        type: string
      native_place:
        type: string
      phone_number:
        type: string
      political_status:
        type: integer
      profile_pic:
        type: string
      resume_id:
        type: string
      work_experience:
        type: string
    required:
    - birthday
    - email
    - name
    - phone_number
    type: object
  request.Resume:
    properties:
      resume_name:
        type: string
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data:
        description: omitempty nil or default
        type: object
      msg:
        type: string
    type: object
host: localhost:8000
info:
  contact:
    email: akazwz@pm.me
    name: API Support
    url: https://akazwz.com
  description: Golang Resume ASpi
  license:
    name: MIT
    url: https://mit-license.org/
  termsOfService: https://akazwz.com
  title: Golang Resume Api
  version: "1.0"
paths:
  /basic-infos:
    post:
      consumes:
      - application/json
      description: 新建简历基本信息
      parameters:
      - description: json
        in: body
        name: basicInfo
        required: true
        schema:
          $ref: '#/definitions/request.BasicInfo'
      - description: Language
        in: query
        name: lang
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.BasicInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 新建基本信息
      tags:
      - basicInfo
  /basic-infos/{resume_id}:
    delete:
      description: 删除基本信息
      parameters:
      - description: Resume ID
        in: path
        name: resume_id
        required: true
        type: string
      - description: Language
        in: query
        name: lang
        type: string
      responses:
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除基本信息
      tags:
      - basicInfo
  /resumes:
    get:
      description: 获取所有简历
      parameters:
      - description: Language
        in: query
        name: lang
        type: string
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取所有简历
      tags:
      - resumes
    post:
      consumes:
      - application/json
      description: 新建简历
      parameters:
      - description: json
        in: body
        name: Resume
        required: true
        schema:
          $ref: '#/definitions/request.Resume'
      - description: Language
        in: query
        name: lang
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Resume'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 新建简历
      tags:
      - resumes
  /resumes/{resume_id}:
    delete:
      description: 删除简历
      parameters:
      - description: Resume ID
        in: path
        name: resume_id
        required: true
        type: string
      - description: Language
        in: query
        name: lang
        type: string
      responses:
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除简历
      tags:
      - resumes
    put:
      description: 更新简历
      parameters:
      - description: Resume ID
        in: path
        name: resume_id
        required: true
        type: string
      - description: json
        in: body
        name: Resume
        required: true
        schema:
          $ref: '#/definitions/request.Resume'
      - description: Language
        in: query
        name: lang
        type: string
      responses:
        "201":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新简历
      tags:
      - resumes
swagger: "2.0"